feat: run tests in parallel when using the same move source
