+++ crates/move-to-polka/tests/examples_basic_test.rs
@@ -61,137 +61,6 @@ pub fn test_error() -> anyhow::Result<()> {
 
 #[test]
 #[serial]
-pub fn test_serialize_signer() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/serialize.polkavm",
-        "../../examples/basic/sources/serialize.move",
-        vec![],
-    )?;
-    let mut address_bytes = [1u8; ACCOUNT_ADDRESS_LENGTH];
-    // set markers for debug displaying
-    address_bytes[0] = 0xab;
-    address_bytes[ACCOUNT_ADDRESS_LENGTH - 1] = 0xce;
-
-    let move_signer = MoveSigner(MoveAddress(address_bytes));
-
-    let signer_address = allocator.copy_to_guest(&mut instance, &move_signer)?;
-
-    instance
-        .call_typed_and_get_result::<u32, (u32,)>(&mut allocator, "ser_signer", (signer_address,))
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-
-    Ok(())
-}
-
-#[test]
-#[serial]
-pub fn test_rv_bool() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/returns.polkavm",
-        "../../examples/basic/sources/returns.move",
-        vec![],
-    )?;
-    let result = instance
-        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_bool_false", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-    assert_eq!(result, 0);
-    let result = instance
-        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_bool_true", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-    assert_eq!(result, 1);
-
-    Ok(())
-}
-
-#[test]
-#[serial]
-pub fn test_rv_u16() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/returns.polkavm",
-        "../../examples/basic/sources/returns.move",
-        vec![],
-    )?;
-    let result = instance
-        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_u16", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-    assert_eq!(result, 19);
-
-    Ok(())
-}
-
-#[test]
-#[serial]
-pub fn test_rv_u32() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/returns.polkavm",
-        "../../examples/basic/sources/returns.move",
-        vec![],
-    )?;
-    let result = instance
-        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_u32", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-    assert_eq!(result, 19);
-
-    Ok(())
-}
-
-#[test]
-#[serial]
-pub fn test_rv_u8() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/returns.polkavm",
-        "../../examples/basic/sources/returns.move",
-        vec![],
-    )?;
-    let result = instance
-        .call_typed_and_get_result::<i32, ()>(&mut allocator, "rv_u8", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-    assert_eq!(result, 19);
-
-    Ok(())
-}
-
-#[test]
-#[serial]
-pub fn test_sha2() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/hash_tests.polkavm",
-        "../../examples/hash_tests/sources/hash_tests.move",
-        vec![],
-    )?;
-    let result = instance
-        .call_typed_and_get_result::<(), ()>(&mut allocator, "sha2_256_expected_hash", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"));
-    assert!(result.is_ok());
-
-    Ok(())
-}
-
-#[test]
-#[serial]
-pub fn test_sha3() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/hash_tests.polkavm",
-        "../../examples/hash_tests/sources/hash_tests.move",
-        vec![],
-    )?;
-    let result = instance
-        .call_typed_and_get_result::<(), ()>(&mut allocator, "sha3_256_expected_hash", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"));
-    assert!(result.is_ok());
-
-    Ok(())
-}
-
-#[test]
-#[serial]
 pub fn test_arith() -> anyhow::Result<()> {
     initialize_logger();
     let (mut instance, mut allocator) = new_move_program(
