+++ crates/move-to-polka/tests/returns.rs
@@ -12,66 +12,62 @@ fn create_blob_once() -> ProgramBlob {
         .get_or_init(|| {
             initialize_logger();
             create_blob(
-                "output/string.polkavm",
-                "../../examples/basic/sources/string.move",
+                "output/returns.polkavm",
+                "../../examples/basic/sources/returns.move",
                 vec![],
             )
             .expect("Failed to compile Move source to PolkaVM bytecode")
         })
         .clone()
 }
-
 #[test]
-pub fn test_string() -> anyhow::Result<()> {
+pub fn test_rv_bool() -> anyhow::Result<()> {
     let blob = create_blob_once();
     let (mut instance, mut allocator) = create_instance(blob)?;
-    instance
-        .call_typed_and_get_result::<(), ()>(&mut allocator, "foo", ())
+    let result = instance
+        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_bool_false", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+    assert_eq!(result, 0);
+    let result = instance
+        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_bool_true", ())
+        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+    assert_eq!(result, 1);
 
     Ok(())
 }
 
 #[test]
-pub fn test_string_index_of() -> anyhow::Result<()> {
+pub fn test_rv_u16() -> anyhow::Result<()> {
     let blob = create_blob_once();
     let (mut instance, mut allocator) = create_instance(blob)?;
-    instance
-        .call_typed_and_get_result::<u64, ()>(&mut allocator, "index_of", ())
+    let result = instance
+        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_u16", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+    assert_eq!(result, 19);
 
     Ok(())
 }
 
 #[test]
-pub fn test_string_substring() -> anyhow::Result<()> {
+pub fn test_rv_u32() -> anyhow::Result<()> {
     let blob = create_blob_once();
     let (mut instance, mut allocator) = create_instance(blob)?;
-    instance
-        .call_typed_and_get_result::<(), ()>(&mut allocator, "substring", ())
+    let result = instance
+        .call_typed_and_get_result::<u32, ()>(&mut allocator, "rv_u32", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+    assert_eq!(result, 19);
 
     Ok(())
 }
 
 #[test]
-pub fn test_append() -> anyhow::Result<()> {
+pub fn test_rv_u8() -> anyhow::Result<()> {
     let blob = create_blob_once();
     let (mut instance, mut allocator) = create_instance(blob)?;
-    instance
-        .call_typed_and_get_result::<u32, ()>(&mut allocator, "append", ())
+    let result = instance
+        .call_typed_and_get_result::<i32, ()>(&mut allocator, "rv_u8", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
-
-    Ok(())
-}
-
-#[test]
-pub fn test_insert() -> anyhow::Result<()> {
-    let blob = create_blob_once();
-    let (mut instance, mut allocator) = create_instance(blob)?;
-    instance
-        .call_typed_and_get_result::<u32, ()>(&mut allocator, "insert", ())
-        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+    assert_eq!(result, 19);
 
     Ok(())
 }
