+++ crates/move-to-polka/tests/serialize.rs
@@ -1,0 +1,54 @@
+use move_to_polka::{
+    initialize_logger,
+    linker::{create_blob, create_instance},
+};
+use once_cell::sync::OnceCell;
+use polkavm::ProgramBlob;
+use polkavm_move_native::types::{MoveAddress, MoveSigner, ACCOUNT_ADDRESS_LENGTH};
+
+static COMPILE_ONCE: OnceCell<ProgramBlob> = OnceCell::new();
+
+fn create_blob_once() -> ProgramBlob {
+    COMPILE_ONCE
+        .get_or_init(|| {
+            initialize_logger();
+            create_blob(
+                "output/serialize.polkavm",
+                "../../examples/basic/sources/serialize.move",
+                vec![],
+            )
+            .expect("Failed to compile Move source to PolkaVM bytecode")
+        })
+        .clone()
+}
+
+#[test]
+pub fn test_serialize_string() -> anyhow::Result<()> {
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
+    instance
+        .call_typed_and_get_result::<u32, ()>(&mut allocator, "ser_string", ())
+        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+
+    Ok(())
+}
+
+#[test]
+pub fn test_serialize_signer() -> anyhow::Result<()> {
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
+    let mut address_bytes = [1u8; ACCOUNT_ADDRESS_LENGTH];
+    // set markers for debug displaying
+    address_bytes[0] = 0xab;
+    address_bytes[ACCOUNT_ADDRESS_LENGTH - 1] = 0xce;
+
+    let move_signer = MoveSigner(MoveAddress(address_bytes));
+
+    let signer_address = allocator.copy_to_guest(&mut instance, &move_signer)?;
+
+    instance
+        .call_typed_and_get_result::<u32, (u32,)>(&mut allocator, "ser_signer", (signer_address,))
+        .map_err(|e| anyhow::anyhow!("{e:?}"))?;
+
+    Ok(())
+}
