+++ crates/move-to-polka/tests/vector.rs
@@ -1,15 +1,30 @@
-use move_to_polka::{initialize_logger, linker::new_move_program};
-use serial_test::serial;
+use move_to_polka::{
+    initialize_logger,
+    linker::{create_blob, create_instance},
+};
+use once_cell::sync::OnceCell;
+use polkavm::ProgramBlob;
+
+static COMPILE_ONCE: OnceCell<ProgramBlob> = OnceCell::new();
+
+fn create_blob_once() -> ProgramBlob {
+    COMPILE_ONCE
+        .get_or_init(|| {
+            initialize_logger();
+            create_blob(
+                "output/vector.polkavm",
+                "../../examples/basic/sources/vector.move",
+                vec![],
+            )
+            .expect("Failed to compile Move source to PolkaVM bytecode")
+        })
+        .clone()
+}
 
 #[test]
-#[serial]
 pub fn test_vector_new() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u64, ()>(&mut allocator, "vecnew", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -18,14 +33,9 @@ pub fn test_vector_new() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_isempty() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<i32, ()>(&mut allocator, "vecisempty", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -34,14 +44,9 @@ pub fn test_vector_isempty() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_cmp() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "veccmp", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -50,14 +55,9 @@ pub fn test_vector_cmp() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_singleton() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "singleton", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -66,14 +66,9 @@ pub fn test_vector_singleton() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_popback() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "popback", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -82,14 +77,9 @@ pub fn test_vector_popback() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_reverse() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "reverse", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -98,14 +88,9 @@ pub fn test_vector_reverse() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_contains() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "contains", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -114,14 +99,9 @@ pub fn test_vector_contains() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_swapremove() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "swapremove", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -130,14 +110,9 @@ pub fn test_vector_swapremove() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_remove() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "remove", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -146,14 +121,9 @@ pub fn test_vector_remove() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_indexof() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "indexof", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -162,14 +132,9 @@ pub fn test_vector_indexof() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_foreach() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "foreach", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -178,14 +143,9 @@ pub fn test_vector_foreach() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_foreachref() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "foreachref", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -194,14 +154,9 @@ pub fn test_vector_foreachref() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_fold() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "fold", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -210,14 +165,9 @@ pub fn test_vector_fold() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_map() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "map", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
@@ -226,14 +176,9 @@ pub fn test_vector_map() -> anyhow::Result<()> {
 }
 
 #[test]
-#[serial]
 pub fn test_vector_filter() -> anyhow::Result<()> {
-    initialize_logger();
-    let (mut instance, mut allocator) = new_move_program(
-        "output/vector.polkavm",
-        "../../examples/basic/sources/vector.move",
-        vec![],
-    )?;
+    let blob = create_blob_once();
+    let (mut instance, mut allocator) = create_instance(blob)?;
     instance
         .call_typed_and_get_result::<u32, ()>(&mut allocator, "filter", ())
         .map_err(|e| anyhow::anyhow!("{e:?}"))?;
